# results/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/models/auto-initiate": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Auto Initiate",
        "description": "Create a dataset and train its top models in the background.",
        "operationId": "auto_initiate_models_auto_initiate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/auto-inference": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Run Auto Inference",
        "description": "Run inference using a dataset's top model.",
        "operationId": "run_auto_inference_models_auto_inference_post",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/inference": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Run Inference",
        "description": "Run inference on a dataset with an explicit input payload.",
        "operationId": "run_inference_models_inference_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/anomaly/auto": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Run Auto Anomaly",
        "description": "Train an anomaly model and start its worker in the background.",
        "operationId": "run_auto_anomaly_models_anomaly_auto_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnomalyRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/clusters": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Clusters",
        "description": "Get clustering results for a dataset.",
        "operationId": "get_clusters_models__dataset_name__clusters_get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/forecast": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Forecast",
        "description": "Get a time-series forecast, refreshing the model if it is expired.",
        "operationId": "get_forecast_models__dataset_name__forecast_get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/forecast/get_actual": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Refresh Forecast Actuals",
        "description": "Refresh a time-series dataset's actual values from InfluxDB.",
        "operationId": "refresh_forecast_actuals_models__dataset_name__forecast_get_actual_get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/find-top": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Find Top Models",
        "description": "Search for the best-performing models in the background.",
        "operationId": "find_top_models_models_find_top_post",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/train": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Train Model",
        "description": "Train a single model in the background.",
        "operationId": "train_model_models_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelTrainSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/change-top/{algorithm}": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Change Top Model",
        "description": "Set a dataset's top model to the given algorithm.",
        "operationId": "change_top_model_models__dataset_name__change_top__algorithm__post",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          },
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Algorithm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/models/algorithms/{task_type}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Algorithms",
        "description": "List available algorithms for a task type.",
        "operationId": "list_algorithms_models_algorithms__task_type__get",
        "parameters": [
          {
            "name": "task_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/by-dataset/{dataset_name}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models By Dataset",
        "description": "List all models belonging to a dataset.",
        "operationId": "list_models_by_dataset_models_by_dataset__dataset_name__get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Models",
        "description": "Delete all models belonging to a dataset.",
        "operationId": "delete_models_models_by_dataset__dataset_name__delete",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/history/anomaly": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get All Anomaly History",
        "description": "Get anomaly prediction history across all datasets.",
        "operationId": "get_all_anomaly_history_models_history_anomaly_get",
        "parameters": [
          {
            "name": "n_day",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "N Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/history": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Prediction History",
        "description": "Get a dataset's prediction history (from InfluxDB, or clusters for clustering).",
        "operationId": "get_prediction_history_models__dataset_name__history_get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/history/anomaly": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Anomaly History",
        "description": "Get a dataset's anomaly prediction history.",
        "operationId": "get_anomaly_history_models__dataset_name__history_anomaly_get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          },
          {
            "name": "n_day",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "N Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_name}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "description": "Get a model by name.",
        "operationId": "get_model_models__model_name__get",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{dataset_name}/results": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Clean Model Results",
        "description": "Clean a dataset's inference results.",
        "operationId": "clean_model_results_models__dataset_name__results_delete",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDatasetSchema": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "features": {
            "items": {},
            "type": "array",
            "title": "Features"
          },
          "tags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Target"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "time_start": {
            "type": "string",
            "title": "Time Start",
            "default": "00:00:00"
          },
          "time_end": {
            "type": "string",
            "title": "Time End",
            "default": "23:59:00"
          },
          "interval": {
            "type": "integer",
            "title": "Interval",
            "default": 0
          },
          "preprocessing": {
            "$ref": "#/components/schemas/PreprocessingSchema"
          },
          "n_models": {
            "type": "integer",
            "title": "N Models",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "features",
          "target",
          "start_date",
          "end_date"
        ],
        "title": "CreateDatasetSchema"
      },
...
    }
  }
}